Migrate from coreferee to fastcoref
coreferee requires spacy<3.6, which has some compatibility
issues with our langchain dependencies that rely on Pydantic
v2. Instead, switch to fastcoref and refactor the NLP
interface.

Instead of providing each NLP pipeline component as a
separate function, we now just surface a wrapper around the
pipeline Doc result. This means we no longer need multiple
pipeline runs to retrieve values, and changes how we handle
coreference resolution.

Previously, we only extracted named entities present in the
message, but for now we'll keep it simple and extract all
entities including those in the context.

Coreference resolution is simplified to just storing the
resolved text as the memory.